feat: 경쟁률 페이지 SEO 개선
- Server Component와 Client Component 분리
- generateMetadata로 동적 메타데이터 생성 (학교명, 시즌 포함)
- heading 구조 개선 (h2 → h1, "실시간 경쟁률" 키워드 추가)
- 검색엔진 최적화를 위한 title, description, keywords 설정
- Open Graph 및 Twitter 카드 메타데이터 추가
